Delete computer data only after 30 days

+++ server/app/Console/Kernel.php
@@ -46,10 +46,13 @@ class Kernel extends ConsoleKernel
         })->daily();
         $schedule->call(function () {
             foreach (Computer::onlyTrashed()->get() as $computer) {
-                $path = 'b2:' . env('B2_BUCKET_NAME') . '/' . $computer->uuid;
-                $process = new Process(['rclone', 'purge', $path]);
-                $process->run();
-                $computer->forceDelete();
+                $delta = $computer->deleted_at->diff(new \DateTime());
+                if ($delta->days > 30) {
+                    $path = 'b2:' . env('B2_BUCKET_NAME') . '/' . $computer->uuid;
+                    $process = new Process(['rclone', 'purge', $path]);
+                    $process->run();
+                    $computer->forceDelete();
+                }
             }
         })->everyMinute();
     }
